Add PMTAdBlocker
A simple chrome extension that blocks adverts on PMT's PDF viewer. PMT (Physics and Maths Tutor - a revision website for 15-19yos in the UK) likes to put a sidebar of adverts on their PDFs, with videos that half the time aren't even targeted - GCSE Chemistry ads on an A-Level Further Maths paper? Doesn't make sense. This extension removes the ads and the sidebar, so you can enjoy the PDFs in full size!